requirements.txt
openai
python-dotenv
yfinance
packaging
duckduckgo-search
fastapi
uvicorn
groq
phidata
sqlalchemy
pgvector
psycopg[binary]
pypdf
sentence-transformers